feat: 보유 자산 요약 API 구현

+++ backend/funco-asset/src/main/java/com/found_404/funco/asset/dto/HoldingCoinInfo.java
@@ -4,8 +4,8 @@ import lombok.Builder;
 
 @Builder
 public record HoldingCoinInfo(
-        String ticker,
-        Double volume,
-        Long averagePrice
+	String ticker,
+	Double volume,
+	Double averagePrice
 ) {
 }
